refactor: Transform seeder to abstract class
