+++ sample/seeds/create-pets.seed.ts
@@ -1,9 +1,10 @@
 import { Connection } from 'typeorm'
-import { Seeder, Factory } from '../../src/types'
+import { Seeder } from '../../src/seeder'
+import { Factory } from '../../src/types'
 import { Pet } from '../entities/Pet.entity'
 
-export default class CreatePets implements Seeder {
-  public async run(factory: Factory, connection: Connection): Promise<any> {
+export default class CreatePets extends Seeder {
+  public async run(factory: Factory, connection: Connection): Promise<void> {
     await factory(Pet)().create()
   }
 }
