+++ src/seeder.ts
@@ -1,0 +1,6 @@
+import { Connection } from 'typeorm'
+import { Factory } from './types'
+
+export abstract class Seeder {
+  abstract run(factory: Factory, connection: Connection): Promise<void>
+}
